refactor: import 추상화 컨벤션 적용
