+++ src/components/common/Icon/CloseIcon.tsx
@@ -1,5 +1,5 @@
 import { Close } from '@/assets';
-import { cn } from '@/lib/utils';
+import { cn } from '@/lib';
 
 type Props = {
   className?: string;
